requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
databases
python-jose[cryptography]
passlib[bcrypt]
python-multipart